modify homepage UI

+++ MakeYourDay/Model/MissionDataModel.swift
@@ -11,8 +11,8 @@ struct MissionData: Codable {
     var activity: String = ""
     var type: String = ""
     var participants: Int = 0
-    var price: Int = 0
+    var price: Double = 0.0
     var link: String = ""
     var key: String = ""
-    var accessibility: Int = 0
+    var accessibility: Double = 0.0
 }
